Java Collections Framework: List. Homework. Additionally, I refactored the project using StringBuilder
